requirements.txt
youtube_transcript_api
streamlit
Google-generativeai
python-dotenv
pathlib